finish the rank module
完成排名信息的记录

+++ AWTennis/src/com/awsports/controller/ActivityController.java
@@ -14,6 +14,7 @@ import com.awsports.pojo.ActivityQuery;
 import com.awsports.pojo.Arena;
 import com.awsports.service.ActivityService;
 import com.awsports.service.ArenaService;
+import com.awsports.util.CustomException;
 
 @Controller
 @RequestMapping("/activity")
@@ -24,6 +25,17 @@ public class ActivityController {
 	@Autowired
 	private ArenaService arenaService;
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:17:37
+	 * @param model
+	 * @param activityQuery
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 活动信息列表
+	 *
+	 */
 	@RequestMapping("/list")
 	public String list(Model model, ActivityQuery activityQuery) throws Exception{
 		List<ActivityQuery> activities=activityService.findAll(activityQuery);
@@ -31,6 +43,17 @@ public class ActivityController {
 		return "activity/list";
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:17:53
+	 * @param model
+	 * @param id
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 更新或添加活动
+	 *
+	 */
 	@RequestMapping("/update")
 	public String update(Model model, Integer id) throws Exception{
 		Activity activity=null;
@@ -45,6 +68,18 @@ public class ActivityController {
 		return "activity/update";
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:18:07
+	 * @param activity
+	 * @param br
+	 * @param model
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 保存修改后的信息
+	 *
+	 */
 	@RequestMapping("/save")
 	public String save(@Validated Activity activity, BindingResult br, Model model) throws Exception{
 		if(br.hasErrors()){
@@ -60,9 +95,23 @@ public class ActivityController {
 		}
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:18:30
+	 * @param id
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 删除活动
+	 *
+	 */
 	@RequestMapping("/delete")
 	public String delete(Integer id) throws Exception{
-		activityService.deleteById(id);
-		return "redirect:list";
+		if(id==null){
+			throw new CustomException("非法操作！");
+		}else{
+			activityService.deleteById(id);
+			return "redirect:list";
+		}
 	}
 }
