+++ AWTennis/src/com/awsports/controller/ArenaController.java
@@ -21,6 +21,17 @@ public class ArenaController {
 	@Autowired
 	private ArenaService arenaService;
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:16:12
+	 * @param model
+	 * @param arena
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 场馆信息列表
+	 *
+	 */
 	@RequestMapping("/list")
 	public String list(Model model, Arena arena) throws Exception{
 		List<Arena> arenas=arenaService.findAll(arena);		
@@ -30,6 +41,17 @@ public class ArenaController {
 		return "arena/list";
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:16:34
+	 * @param model
+	 * @param id
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 更新或添加场馆
+	 *
+	 */
 	@RequestMapping("/update")
 	public String update(Model model, Integer id) throws Exception{
 		Arena arena=null;
@@ -43,6 +65,18 @@ public class ArenaController {
 		return "arena/update";
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:16:51
+	 * @param arena
+	 * @param br
+	 * @param model
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 保存修改后的场馆信息
+	 *
+	 */
 	@RequestMapping("/save")
 	public String save(@Validated Arena arena, BindingResult br, Model model) throws Exception{
 		if(br.hasErrors()){
@@ -66,10 +100,24 @@ public class ArenaController {
 		}
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:17:14
+	 * @param id
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 删除场馆记录
+	 *
+	 */
 	@RequestMapping("/delete")
 	public String delete(Integer id) throws Exception{
-		arenaService.deleteById(id);
-		//删除成功，跳转到列表页面
-		return "redirect:list";
+		if(id==null){
+			throw new CustomException("非法操作！");
+		}else{
+			arenaService.deleteById(id);
+			//删除成功，跳转到列表页面
+			return "redirect:list";
+		}
 	}
 }
