+++ AWTennis/src/com/awsports/controller/PointruleController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.awsports.pojo.Pointrule;
 import com.awsports.service.PointruleService;
+import com.awsports.util.CustomException;
 import com.awsports.util.TypeMap;
 
 @Controller
@@ -18,6 +19,18 @@ public class PointruleController {
 	@Autowired
 	private PointruleService pointruleService;
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:20:42
+	 * @param model
+	 * @param id
+	 * @param tournamentid
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 更新或添加积分规则
+	 *
+	 */
 	@RequestMapping("/update")
 	public String update(Model model,Integer id,Integer tournamentid) throws Exception{
 		Pointrule pointrule=null;
@@ -32,8 +45,18 @@ public class PointruleController {
 		return "pointrule/update";
 	}
 	
-	
-	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:22:12
+	 * @param pointrule
+	 * @param br
+	 * @param model
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 保存修改后的记录 
+	 *
+	 */
 	@RequestMapping("/save")
 	public String save(@Validated Pointrule pointrule, BindingResult br, Model model) throws Exception{
 		if(br.hasErrors()){
@@ -51,11 +74,27 @@ public class PointruleController {
 		}
 	}
 	
+	/**
+	 * 
+	 * @Author: peRFect
+	 * @Datetime: 2017年6月26日 下午6:22:43
+	 * @param model
+	 * @param id
+	 * @param tournamentid
+	 * @throws Exception
+	 * @Return: String
+	 * @Description: 删除记录
+	 *
+	 */
 	@RequestMapping("/delete")
 	public String delete(Model model,Integer id,Integer tournamentid) throws Exception{
-		pointruleService.deleteById(id);
-		model.addAttribute("id", tournamentid);
-		return "redirect:/tournament/detail";
+		if(id==null){
+			throw new CustomException("非法操作！");
+		}else{
+			pointruleService.deleteById(id);
+			model.addAttribute("id", tournamentid);
+			return "redirect:/tournament/detail";
+		}
 	}
 	
 }
